+++ LinqToTwitter/SocialGraph.cs
@@ -18,13 +18,13 @@ namespace LinqToTwitter
         /// <summary>
         /// The ID or screen_name of the user to retrieve the friends ID list for
         /// </summary>
-        public uint ID { get; set; }
+        public ulong ID { get; set; }
 
         /// <summary>
         /// Specfies the ID of the user for whom to return the friends list. 
         /// Helpful for disambiguating when a valid user ID is also a valid screen name. 
         /// </summary>
-        public uint UserID { get; set; }
+        public ulong UserID { get; set; }
 
         /// <summary>
         /// Specfies the screen name of the user for whom to return the friends list.
